capstone2-server/requirements.txt
requests>=2.32.3,<3.0.0
django>=5.1.4,<6.0.0
djangorestframework>=3.15.2,<4.0.0
gunicorn>=23.0.0,<24.0.0
psycopg2-binary>=2.9.10,<3.0.0
django-stubs>=5.1.1,<6.0.0
pyjwt>=2.10.1,<3.0.0
django-filter>=24.3,<25.0
